Split up the 1000-line behemoth that was PianoRoll.tsx into multiple components and hooks
